==> 3-clustering_and_annotation/02.3-cluster_365.Rmd <==
---
title: 'Cluster tumoral cells (case 365)'
author: "Ramon Massoni-Badosa"
date: "2021/04/09"
output: 
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "100%", fig.align='center', 
                      message=FALSE, warning = FALSE)
options(width = 1200)
```


# Introduction

Here, we will cluster the cells of the patient with id "365".


# Pre-processing

## Load packages

```{r}
library(Seurat)
library(SeuratWrappers)
library(harmony)
library(tidyverse)
```


## Define parameters

```{r}
# Paths
path_to_obj <- here::here("results/R_objects/4.seurat_leukemic.rds")
path_to_save <- here::here("results/R_objects/5.seurat_clustered_365.rds")


# Colors
color_palette <- c("black", "gray", "red", "yellow", "violet", "green4",
                   "blue", "mediumorchid2", "coral2", "blueviolet",
                   "indianred4", "deepskyblue1", "dimgray", "deeppink1",
                   "green", "lightgray", "hotpink1")


# Source functions
source(here::here("bin/utils.R"))


# Thresholds
```

  
## Load data

```{r}
seurat <- readRDS(path_to_obj)
```


# Subset and reprocess

Of note, everytime that we filter out a specific population, the set of highly variable genes (HVG) and the axis of variability (PCs) changes. Thus, we need to rerun the main analysis steps iteratively.

First, we focus on the patient studied in this notebook:

```{r}
seurat <- subset(seurat, donor_id == "365")
seurat <- process_seurat(seurat, dims = 1:20)
DimPlot(seurat, group.by = "subproject")
```

As we can see, there is a large batch effect between the two experiments we conducted BCLLATLAS_10 and BCLLATLAS_29. However, contrary to the previous cases, here BCLLATLAS_29 contains all time-points and has a better quality than BCLLATLAS_10. Thus, in this case we will focus first on BCLLATLAS_29 (not hashed) and later in BCLLATLAS_10 (hashed).


```{r}
seurat <- subset(seurat, subproject == "BCLLATLAS_29")
seurat <- process_seurat(seurat, dims = 1:20)
DimPlot(seurat, group.by = "subproject")
FeaturePlot(seurat, "pct_mt") +
  scale_color_viridis_c(option = "magma")
FeaturePlot(seurat, "nFeature_RNA") +
  scale_color_viridis_c(option = "magma")
```


We see a subpopulation with characteristics of low-quality cells: high mitochondrial expression and low number of detected features (genes). Hence, let us exclude it from the dataset:

```{r}
seurat <- FindNeighbors(seurat, reduction = "pca", dims = 1:20)
seurat <- FindClusters(seurat, resolution = 0.3)
DimPlot(seurat)
seurat <- subset(seurat, seurat_clusters != "2")
seurat <- process_seurat(seurat, dims = 1:20)
DimPlot(seurat)
DimPlot(seurat, group.by = "time_point")
FeaturePlot(seurat, features = "NKG7")
```

Despite our efforts, we still have some cells from the microenvironment:

```{r}
# microenv_barcodes <- CellSelector(FeaturePlot(seurat, features = "NKG7"))
microenv_barcodes <- c(
  "s6dm1ahu_e8w2ucdi_AAGTGAACAGGTCCGT-1",
  "s6dm1ahu_e8w2ucdi_CGGGACTCACCTTCGT-1",
  "s6dm1ahu_e8w2ucdi_GAAGGGTGTAGTGATA-1",
  "s6dm1ahu_e8w2ucdi_GATCCCTAGGTCCCGT-1",
  "s6dm1ahu_e8w2ucdi_TACAACGGTCCTGAAT-1",
  "s6dm1ahu_e8w2ucdi_TACCCGTAGTAGTCAA-1"
)
seurat$is_microenv <- ifelse(
  colnames(seurat) %in% microenv_barcodes,
  TRUE,
  FALSE
)
Idents(seurat) <- "is_microenv"
markers_microenv <- FindMarkers(
  seurat,
  ident.1 = TRUE,
  ident.2 = FALSE,
  only.pos = TRUE,
  logfc.threshold = 0.75
)
DT::datatable(markers_microenv)
```


Filter out:

```{r}
seurat <- subset(seurat, is_microenv == FALSE)
```


# Cluster

```{r}
seurat <- FindNeighbors(seurat, reduction = "pca", dims = 1:20)
seurat <- FindClusters(seurat, resolution = 0.1)
DimPlot(seurat)


# Find subclusters Richter
seurat <- FindSubCluster(
  seurat,
  cluster = "1",
  subcluster.name = "richter_subclusters",
  graph.name = "RNA_snn",
  resolution = 0.3
)
DimPlot(seurat, group.by = "richter_subclusters")


# Find subclusters CLL
Idents(seurat) <- "richter_subclusters"
seurat <- FindSubCluster(
  seurat,
  cluster = "0",
  subcluster.name = "cll_subclusters",
  graph.name = "RNA_snn",
  resolution = 0.2
)
DimPlot(seurat, group.by = "cll_subclusters")
Idents(seurat) <- "cll_subclusters"
markers_0 <- FindMarkers(
  seurat,
  ident.1 = "0_1",
  ident.2 = "0_0",
  only.pos = FALSE
)
markers_0 <- markers_0 %>%
  rownames_to_column("gene") %>%
  arrange(desc(avg_log2FC))
DT::datatable(markers_0)
```


Importantly we see that in the CLL cluster the clusters are not very robust, since the fold-changes are very small. Thus, we will consider it as a single cluster:


```{r}
seurat$final_clusters <- seurat$richter_subclusters
Idents(seurat) <- "final_clusters"
DimPlot(seurat)
```


# Save

```{r}
saveRDS(seurat, path_to_save)
```


# Session Info

```{r}
sessionInfo()
```
