==> 3-clustering_and_annotation/03.3-annotate_365.Rmd <==
---
title: 'Annotate tumoral cells (case 365)'
author: "Ramon Massoni-Badosa"
date: "2021/04/09"
output: 
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "100%", fig.align='center', 
                      message=FALSE, warning = FALSE)
options(width = 1200)
```


# Introduction

Here, we will annotate the cells of the patient with id "365".


# Pre-processing

## Load packages

```{r}
library(Seurat)
library(SeuratWrappers)
library(harmony)
library(tidyverse)
```


## Define parameters

```{r}
# Paths
path_to_obj <- here::here("results/R_objects/5.seurat_clustered_365.rds")
path_to_obj_sampl_artifacts <- here::here("results/R_objects/cll_seurat_annotated.rds")
path_to_save <- here::here("results/R_objects/6.seurat_annotated_365.rds")
path_to_save_markers <- here::here("3-clustering_and_annotation/markers_clusters_365.rds")


# Colors
color_palette <- c("black", "gray", "red", "yellow", "violet", "green4",
                   "blue", "mediumorchid2", "coral2", "blueviolet",
                   "indianred4", "deepskyblue1", "dimgray", "deeppink1",
                   "green", "lightgray", "hotpink1")


# Source functions
source(here::here("bin/utils.R"))


# Thresholds
min_log2FC <- 0.3
alpha <- 0.001
```

  
## Load data

```{r}
seurat <- readRDS(path_to_obj)
seurat
DimPlot(seurat, cols = color_palette)
seurat_bias <- readRDS(path_to_obj_sampl_artifacts)
```


# Cell Cycle Score

As upregulation of cell cycle genes is a hallmark of Richter transformation, we will infer the cell cycle score and phase for each cell:

```{r}
DimPlot(seurat, group.by = "time_point")
FeaturePlot(seurat, "CD3D")

seurat <- CellCycleScoring(
  seurat,
  s.features = cc.genes.updated.2019$s.genes,
  g2m.features = cc.genes.updated.2019$g2m.genes,
  set.ident = FALSE
)
DimPlot(seurat, group.by = "Phase")
umap_s_score <- FeaturePlot(seurat, features = "S.Score") +
  scale_color_viridis_c(option = "magma") +
  labs(title = "S Score") +
  theme(
    plot.title = element_text(hjust = 0.5, size = 12, face = "plain"),
    axis.title = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank(),
    axis.line = element_blank()
  )
umap_g2m_score <- FeaturePlot(seurat, features = "G2M.Score") +
  scale_color_viridis_c(option = "magma") +
  labs(title = "G2M Score") +
  theme(
    plot.title = element_text(hjust = 0.5, size = 12, face = "plain"),
    axis.title = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank(),
    axis.line = element_blank()
  )
umap_cc_combined <- ggpubr::ggarrange(
  plotlist = list(umap_s_score, umap_g2m_score),
  nrow = 2,
  ncol = 1,
  common.legend = FALSE
)
umap_cc_combined
```


# Find Markers

```{r}
markers <- FindAllMarkers(seurat, only.pos = TRUE, logfc.threshold = min_log2FC)
markers <- markers %>%
  mutate(cluster = as.character(cluster)) %>%
  filter(p_val_adj < alpha) %>%
  arrange(cluster) %>%
  group_by(cluster) %>%
  arrange(desc(avg_log2FC), .by_group = TRUE)
DT::datatable(markers, options = list(scrollX = TRUE))
```


# Annotation

Important literature to annotate the cells:


* [Intraclonal Complexity in Chronic Lymphocytic Leukemia:Fractions Enriched in Recently Born/Divided andOlder/Quiescent Cells](https://molmed.biomedcentral.com/articles/10.2119/molmed.2011.00360).
* [MicroRNA-155influences B-cell receptor signaling and associateswith aggressive disease in chronic lymphocytic leukemia](https://ashpublications.org/blood/article/124/4/546/33343/MicroRNA-155-influences-B-cell-receptor-signaling).
* [Of Lymph Nodes and CLL Cells: Deciphering the Role of CCR7 in the Pathogenesis of CLL and Understanding Its Potential as Therapeutic Target](https://www.frontiersin.org/articles/10.3389/fimmu.2021.662866/full)


| Cluster | Markers                             | Annotation                 |
|---------|-------------------------------------|----------------------------|
| 0       | CXCR4                               | CLL                        |
| 1_0     | CD27                                | CXCR4loCD27hi RT           |
| 1_1     | MIR155HG                            | MIR155HGhi RT              |
| 1_2     | ENO1, TCL1A                         | RT quiescent               |
| 1_3     | PCNA, TOP2A                         | RT proliferative           |


Annotate:

```{r}
seurat$annotation_final <- factor(
  seurat$final_clusters,
  levels = c("0", "1_0", "1_1", "1_2", "1_3")
)
new_levels_365 <- c("CLL", "CXCR4loCD27hi RT", "MIR155HGhi RT",
                    "RT quiescent", "RT proliferative")
levels(seurat$annotation_final) <- new_levels_365
seurat$annotation_final <- factor(
  seurat$annotation_final,
  color_annotations$`365`$annotation_final
)
Idents(seurat) <- "annotation_final"


# Plot UMAP
(umap_annotation <- plot_annotation(
  seurat_obj = seurat,
  pt_size = 0.5,
  colors_reference = color_annotations,
  patient_id = "365",
  nothing = TRUE
))
```


# Visualize markers

```{r}
# UMAPs
genes_interest <- c("CXCR4", "CD27", "MIR155HG", "ENO1", "TCL1A",
                    "PCNA", "TOP2A")
(feature_plots <- purrr::map(genes_interest, function(x) {
  p <- FeaturePlot(seurat, x, pt.size = 0.5) +
    scale_color_viridis_c(option = "magma")
  p
}))
 

# Dot plots
(dot_plot <- plot_dot_plot(
  seurat,
  goi = rev(genes_interest),
  colors_reference = color_annotations,
  patient_id = "365"
))


# Violin plots
(vln_plot_s <- plot_violin_plot(
  seurat,
  continuous_var = "S.Score",
  ylab = "S Phase Score",
  colors_reference = color_annotations,
  patient_id = "365"
))
(vln_plot_g2m <- plot_violin_plot(
  seurat,
  continuous_var = "G2M.Score",
  ylab = "G2M Phase Score",
  colors_reference = color_annotations,
  patient_id = "365"
))
```


# Validation 1: CLL-specific clustering

Although we could recover the major clusters that we have observed in previous patients, it is intriguing that the clusters of MIR155HGhi and CXCR4loCD27hi were found in the RT time-point and not in the CLL. Thus, let us perform a CLL-specific clustering:


```{r}
seurat_cll <- subset(seurat, time_point == "T2")
seurat_cll <- process_seurat(seurat_cll, dims = 1:20)
feature_plots2 <- purrr::map(c("CXCR4", "CD27", "MIR155HG", "S100A4"), function(x) {
  p <- FeaturePlot(seurat_cll, x, pt.size = 0.5) +
    scale_color_viridis_c(option = "magma")
  p
})
feature_plots2
```

Indeed, we see that CXCR4, CD27 and MIR155HG are mutually exclusive, but we cannot distinguish 3 clear clusters

```{r}
seurat_cll <- FindNeighbors(seurat_cll, dims = 1:20, reduction = "pca")
seurat_cll <- FindClusters(seurat_cll, resolution = 0.5)
DimPlot(seurat_cll)
```


# Validation 2: sampling artifacts

Could this pattern be explained by a differential processing time between both samples (RT and CLL)?

```{r}
seurat_bias <- subset(seurat_bias, temperature == "RT")
seurat_bias <- subset(seurat_bias, time == "24h")
DimPlot(seurat_bias)
seurat_bias <- SplitObject(seurat_bias, "cell_type")
seurat_bias <- seurat_bias[c("CLL 1472", "CLL 1892")]
seurat_bias <- purrr::map(seurat_bias, process_seurat, dims = 1:20)
DimPlot(seurat_bias$`CLL 1472`, group.by = "time")
FeaturePlot(seurat_bias$`CLL 1472`, c("CXCR4", "CD27", "MIR155HG"))
DimPlot(seurat_bias$`CLL 1892`, group.by = "time")
FeaturePlot(seurat_bias$`CLL 1892`, c("CXCR4", "CD27", "MIR155HG"))
```


# Save

```{r}
# Save Seurat object
saveRDS(seurat, path_to_save)


# Save markers
markers$annotation <- factor(markers$cluster)
levels(markers$annotation) <- new_levels_365
markers_list <- purrr::map(levels(markers$annotation), function(x) {
  df <- markers[markers$annotation == x, ]
  df <- df[, c(7, 1, 5, 2:4, 6, 8)]
  df
})
names(markers_list) <- levels(markers$annotation)
markers_list <- markers_list[color_annotations$`365`$annotation_final]
markers_final <- bind_rows(markers_list)
saveRDS(markers_list, path_to_save_markers)
saveRDS(
  markers_final,
  here::here("results/tables/markers/markers_annotated_clusters_patient_365.rds")
)
saveRDS(markers_list, path_to_save_markers)
openxlsx::write.xlsx(
  x = markers_list,
  file = "results/tables/markers/markers_annotated_clusters_patient_365.xlsx"
)
```


# Session Info

```{r}
sessionInfo()
```
